app/requirements.txt
beautifulsoup4
fastapi
matplotlib
pandas
python-dotenv
pyyaml
requests
uvicorn